# openapi.yaml
openapi: "3.1.2"
info:
  title: Komodo Accounts API
  version: "1.0.0"
  description: |
    Account profile, address, payment method, and preference management.

    **Two servers:**
    - Public (7051): account-facing profile endpoints; require user JWT
    - Internal (7052): service-to-service account lookup endpoints; require service JWT

    **Auth:** All `/v1/me/profile`, `/v1/me/addresses`, `/v1/me/payments`, and `/v1/me/preferences`
    routes require a user-scoped JWT (`Authorization: Bearer <token>`).
    Internal `/v1/accounts/{id}/*` routes require a service-scoped JWT (`svc:` prefix claim).

    **HTTP conventions:**
    - POST → 201 Created / 400 Bad Request / 409 Conflict
    - PUT → 200 OK / 404 Not Found
    - DELETE → 204 No Content / 404 Not Found
    - GET → 200 OK / 404 Not Found
servers:
  - url: http://localhost:7051
    description: Public (local)
  - url: http://localhost:7052
    description: Internal (local)
...
paths:
...
  /v1/me/profile:
    get:
      tags: [Profile]
      summary: Get authenticated account's profile
      operationId: getProfile
      x-visibility: public
      responses:
        "200":
          description: Account profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

    post:
      tags: [Profile]
      summary: Create account record (called on registration)
      operationId: createAccount
      x-visibility: public
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAccountRequest"
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/Conflict"

    put:
      tags: [Profile]
      summary: Update authenticated account's profile
      operationId: updateProfile
      x-visibility: public
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAccountRequest"
      responses:
        "200":
          description: Updated account profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      tags: [Profile]
      summary: Account closure / right-to-delete
      description: Initiates a 30-day soft-delete window; satisfies GDPR right to erasure, CCPA/CPRA right to delete, VCDPA, and CPA. Data is permanently erased after the window expires unless restored.
      operationId: deleteProfile
      x-visibility: public
      responses:
        "202":
          description: Account closure requested
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/me/profile/restore:
    post:
      tags: [Profile]
      summary: Restore account from pending deletion
      description: Cancels a pending account closure if within the 30-day window. Requires account JWT.
      operationId: restoreProfile
      x-visibility: public
      responses:
        "200":
          description: Account restored
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        "409":
          $ref: "#/components/responses/Conflict"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /v1/me/profile/avatar:
    post:
      tags: [Profile]
      summary: Get avatar upload URL
      description: Returns a 15-minute pre-signed S3 PUT URL. Client uploads the image directly to S3, then updates the profile avatar_url field via PUT /v1/me/profile.
      operationId: getAvatarUploadURL
      x-visibility: public
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Pre-signed upload URL
          content:
            application/json:
              schema:
                type: object
                properties:
                  upload_url:
                    type: string
                  expires_in_seconds:
                    type: integer
                    example: 900
        "401":
          $ref: "#/components/responses/Unauthorized"

  # ── Consent ──────────────────────────────────────────────────────────────
...
  /v1/me/addresses:
    get:
      tags: [Addresses]
      summary: List all addresses for the authenticated account
      operationId: getAddresses
      x-visibility: public
      responses:
        "200":
          description: Address list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Address"
        "401":
          $ref: "#/components/responses/Unauthorized"

    post:
      tags: [Addresses]
      summary: Add a new address
      operationId: addAddress
      x-visibility: public
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddressInput"
      responses:
        "201":
          description: Address created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Address"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/me/addresses/{id}:
    parameters:
      - $ref: "#/components/parameters/AddressID"

    put:
      tags: [Addresses]
      summary: Update an address
      operationId: updateAddress
      x-visibility: public
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddressInput"
      responses:
        "200":
          description: Updated address
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Address"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      tags: [Addresses]
      summary: Delete an address
      operationId: deleteAddress
      x-visibility: public
      responses:
        "204":
          description: Address deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

  # ── Payments ─────────────────────────────────────────────────────────────
  /v1/me/payments:
    get:
      tags: [Payments]
      summary: List saved payment methods
      operationId: getPayments
      x-visibility: public
      responses:
        "200":
          description: Payment method list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PaymentMethod"
        "401":
          $ref: "#/components/responses/Unauthorized"

    put:
      tags: [Payments]
      summary: Add or update a payment method
      operationId: upsertPayment
      x-visibility: public
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentMethod"
      responses:
        "200":
          description: Payment method saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentMethod"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/me/payments/{id}:
    parameters:
      - $ref: "#/components/parameters/PaymentID"

    delete:
      tags: [Payments]
      summary: Remove a payment method
      operationId: deletePayment
      x-visibility: public
      responses:
        "204":
          description: Payment method removed
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

  # ── Preferences ──────────────────────────────────────────────────────────
  /v1/me/preferences:
    get:
      tags: [Preferences]
      summary: Get account preferences
      operationId: getPreferences
      x-visibility: public
      responses:
        "200":
          description: Account preferences
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Preferences"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

    put:
      tags: [Preferences]
      summary: Update account preferences
      operationId: updatePreferences
      x-visibility: public
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Preferences"
      responses:
        "200":
          description: Updated preferences
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Preferences"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"

    delete:
      tags: [Preferences]
      summary: Delete account preferences
      operationId: deletePreferences
      x-visibility: public
      responses:
        "204":
          description: Preferences deleted
        "401":
          $ref: "#/components/responses/Unauthorized"

  # ── Internal ─────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    UserJWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: User-scoped RS256 JWT issued by komodo-auth-api
    ServiceJWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Service-scoped RS256 JWT; must carry at least one `svc:` claim
  parameters:
...
    AddressID:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Address ID
    PaymentID:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Payment method ID
...
  schemas:
    Account:
      type: object
      required: [account_id, username, email, first_name, last_name]
      properties:
        account_id:
          type: string
        username:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        first_name:
          type: string
        last_name:
          type: string
        avatar_url:
          type: string
          format: uri
    CreateAccountRequest:
      type: object
      required: [username, email, first_name, last_name]
      properties:
        username:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        first_name:
          type: string
        last_name:
          type: string
        avatar_url:
          type: string
          format: uri
    UpdateAccountRequest:
      type: object
      description: >-
        Only the mutable fields below are applied. Immutable or server-managed
        fields (account_id, username, email, auth_methods, created_at, updated_at)
        are ignored if present in the request body.
      properties:
        phone:
          type: string
        first_name:
          type: string
        last_name:
          type: string
        avatar_url:
          type: string
          format: uri
    Address:
      type: object
      required: [address_id, line1, city, state, zip_code, country, is_default]
      properties:
        address_id:
          type: string
        alias:
          type: string
          description: Friendly label (e.g. "Home", "Work")
        line1:
          type: string
        line2:
          type: string
        city:
          type: string
        state:
          type: string
        zip_code:
          type: string
        country:
          type: string
          description: ISO 3166-1 alpha-2 country code (e.g. "US")
        is_default:
          type: boolean
    AddressInput:
      type: object
      required: [line1, city, state, zip_code, country]
      properties:
        alias:
          type: string
        line1:
          type: string
        line2:
          type: string
        city:
          type: string
        state:
          type: string
        zip_code:
          type: string
        country:
          type: string
          description: ISO 3166-1 alpha-2 country code (e.g. "US")
        is_default:
          type: boolean
    Preferences:
      type: object
      properties:
        language:
          type: string
          description: BCP 47 language tag (e.g. "en-US")
        timezone:
          type: string
          description: IANA timezone (e.g. "America/New_York")
        communication:
          type: object
          description: Notification channel opt-ins keyed by channel name (e.g. email, sms, push)
          additionalProperties:
            type: boolean
    PaymentMethod:
      type: object
      description: Stored payment method reference — no raw card data, metadata only
      required: [payment_id, provider, last4, brand, expiry_month, expiry_year, is_default]
      properties:
        payment_id:
          type: string
        provider:
          type: string
          description: Payment processor (e.g. "stripe")
        last4:
          type: string
          description: Last 4 digits of the card number
        brand:
          type: string
          description: Card network (e.g. "visa", "mastercard", "amex")
        expiry_month:
          type: integer
          minimum: 1
          maximum: 12
        expiry_year:
          type: integer
        is_default:
          type: boolean
...
    ProblemDetail:
      type: object
      description: RFC 7807 Problem+JSON error body
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
  responses:
    BadRequest:
      description: Invalid request body or parameters
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"
    Unauthorized:
      description: Missing or invalid JWT
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"
...
    NotFound:
      description: Resource not found
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"
    Conflict:
      description: Resource already exists
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"
security:
  - UserJWT: []